requirements.txt
africastalking==1.2.3
arrow==1.0.3
asgiref==3.3.1
blessed==1.18.0
certifi==2020.12.5
chardet==4.0.0
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-filter==2.4.0
django-picklefield==3.0.1
django-q==1.3.5
django-rest-swagger==2.2.0
djangorestframework==3.12.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
schema==0.7.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
whitenoise==5.2.0
